HTML file
Created a new HTML file, linked the JS file to it. Added a title. Chose to add both h1 and button without an ID for simplicity reasons, as they are alone in the code and there are no other buttons, for example, to be confused with when referencing in the JS file.