Organising
Just rearranged the code a little better to make thing clearer.

Moved the bit of the function that generates the RPB values to a new function stored in a variable called randomColor. Replaced the last variable for a return.

Within the event function, added a variable called newColor, which is the same as the randomColor variable results.